WIP next_indent priority ordering
1 failing test

The problem is that the these two expansion cases conflict in terms of indentation order:

```
defone
  if value

    print one
  else

    print two
  end
end
```

```
deftwo
  if value
    print one
    print two
   end
end
```

For the defone case you do not want to try to expand from "print two" to grab the "else" until after "print one" has been added to the frontier. Otherwise you end up grabbing just the "else end" and it's invalid and triggers a frontier check. I'm pretty sure we want to fully evaluate all "inner" sections, before expanding to engulf them. I think this logic is currently correct and we shouldn't break it.

For the deftwo case, the end is mismatched with the if. The inner prints are expanded first, then the end is added in the "add" phase and triggers a frontier check. This feels produces an invalid result and needs to be fixed. Intuitively it "feels" like the wrong decision to add rather than expand, but I can't translate that intuition into a hard and fast rule for the algorithm.

## Different algorithm for determining "next" indentation for a line

At the time that "next line" is determined for the "end" in deftwo it's "above" is marked hidden, when in reality it should logically be the "if value" line. In that case it's next indent would be 0 (which is where it's next expansion actually ends up).

Instead of determining next line manually, we could scan to grab empty/hidden lines, and then compare from there.

Seems like it should work, let's try it out