WIP, trying to figure out if tests are broken or my code is
DEBUG=1 be rspec spec/unit/code_search_spec.rb:80


DEBUG=1 be rspec ./spec/integration/dead_end_spec.rb:207

Problem is this:

```
  1  class Dog
  2    def bark
❯ 3      puts "woof"
  4  end
```

Turns into this:

```
    Block lines: 2..4 (expand)

  1  class Dog
❯ 2    def bark
❯ 4  end
```

Which eliminates a a block and then:

```
    Block lines: 1..1 (add)

❯ 1  class Dog
```

I think this can be fixed in post via Capture Context (TBH I thought it already was). I checked and it should be.

SOOOO...there's likely a bug in capture context, maybe there's an extra newline or something in the block and the logic is off a bit.

ALSO

- There are mysterious adds happening that shouldn't be happening in this case. PLZ investigate.