WIP Track Block "parent"
1 failing test 

$ DEAD_END_TIMEOUT=10 be rspec spec/integration/dead_end_spec.rb:8


By tracking what a block was expanded from we get extra data. Specifically when handling blocks that were eagerly expanded (as is done in the Add phase) we can go back and clean them up.

This data is REALLY powerful.

We're fixing one of two failing tests, I like this output better. 

Initially I tried rendering the parent using logic to see if the original block is valid without it. It really zeros in on some problems, for example:

```
          1  module SyntaxErrorSearch
          3    class DisplayInvalidBlocks
       ❯ 36      def filename
       ❯ 38      def code_with_filename
       ❯ 45      end
         63    end
         64  end
```

Becomes:

```
       +   1  module SyntaxErrorSearch
       +   3    class DisplayInvalidBlocks
       +  17      def call
       +  34      end
       +❯ 36      def filename
       +  38      def code_with_filename
       +  45      end
       +  63    end
       +  64  end
```

However it cuts off extra detail that we would otherwise want to preserve, so:

```
         ❯ 1  class Dog
         ❯ 2    def bark
         ❯ 4  end
```

Becomes:

```
       ❯ 1  class Dog
         4  end
```


Which is not great since it doesn't actually show the problem.

There is still potential to clean some of these up to either focus them or unfocus them as we see fit. My initial experiment wasn't perfect, but with more time and effort it could be very valuable.

## It takes 2 (parents)

The other possible "source" a block could come from is when one block engulfs another. Especially since the "Add" phase runs in such a way that the global "holds all syntax errors" check is only triggered after ALL add phases have been run.

I think if we track when one block engulfs another it buys us another vector to check. If a valid block engulfs an invalid block, we should be able to drill back in and find the original invalid block