WIP Initial tree building seems successful
I'm sure there are edge cases but I'm liking the properties of the tree.

It lets me view blocks logically the way I construct them in my head (based on indentation). 


I may want to make some convenience methods, for instance given the if/else/end cadence I want to quickly view



```
          Args.new(parts: [argument], location: argument.location)

          Args.new(
            parts: arguments.parts << argument,
            location: arguments.location.to(argument.location)
          )

```

And:

```
        if arguments.parts.empty?
        else
        end
```

If the problem is in the block, this will let me figure out if I need to "zoom" in another level or if I've already isolated the issue. In the current indentation. 


So far all my tests are mostly with valid code and mostly regular indentation. I don't know how well these properties will hold when we face some "real" world code cases