WIP perf commit
Currently spending 40% of time building the tree in popping off nodes 


```
40.12% 	0.35% 	3.16 	0.03 	0.00 	3.13 23373 	DeadEnd::PriorityQueue#pop
```

Spending 9% of time enqueueing nodes. It might be worth it to revisit using an insertion sort (with bsearch_index) to see if it's faster than a heap implementation as we're guaranteeing to search through all elements in the queue.


Currently perf for tree building our target case is:

```
$ be rspec spec/unit/indent_tree_spec.rb:7 
Run options: include {:locations=>{"./spec/unit/indent_tree_spec.rb"=>[7]}}

DeadEnd::IndentTree
  WIP syntax_tree.rb.txt for performance validation

Finished in 0.75664 seconds (files took 0.14354 seconds to load)
1 example, 0 failures
```

Main has the full search around 1.2~1.1 seconds with the same time to load). So our budget to hit equal perf is 0.45 seconds. I'm hoping we can beat that by a bunch, but we've not even touched ripper yet and it's *supposed* to be the expensive part.