Refactor logic to Diagnose class
Previously the BlockNode knew how to both "diagnose" it's own problem as well
as how to return its parent that likely holds the core issue. That caused
a lot of extra, somewhat confusing code to go into BlockNode that wasn't really
the responsibility of that object.

This commit introduces a new class `Diagnose` it is responsible for naming the problem
with the current node and extracting the likely issue. This makes BlockNode smaller
and more focused, it also isolates all the extraction logic to one convienent location.

This was done as a pure refactor. After fully extracting the existing logic, I was able
to clean up IndentSearch a lot thanks to the new API. It is now cleaner and more focused.
The IndentSearch class is now mainly responsible for maintaining the guarantees that each
journey only contains steps that would produce a valid document.

I also found a problem where we were claiming that a block node was not a leaf even though
it was. This was due to passing in a non-empty parents array to the constructor when we
shouldn't have. That's now been fixed