Fix problem with one bad internal node
When there's a node inside of leaning elements it might look valid in isolation, for example:

```
cat
```

Is valid code but:

```
[
  cat
  dog,
  bird
]
```

Is not (missing a comma). We are already enumerating each internal element to see if it's valid in isolation i.e.:

```
[
  cat
]
```

and

```
[
  dog,
]
```

and

```
[
  bird
]
```

However due to ruby's permissive rules all of these will be true and no single problem will be reported

## The fix

Rather than checking each element in isolation, we can do the opposite. Check if removing that element from the inner block array would produce a valid outcome:

```
[
  cat
  dog,
  # bird
] # Still invalid
```

```
[
  cat
  # dog,
  bird
] # Still invalid
```

```
[
  # cat
  dog,
  bird
] # VALID!
```

If one and only one block can be removed to produce a valid outcome then it must hold the problem.

Worth noting that it feels like there's a more generic case to handle several of the different diagnose states but I'm unable to find it right now. I would love to clean up the Diagnose class to make some of these checks cleaner/simpler