WIP 8 total failing tests
The current block node implementation is optimized for the "hard" cases from the old algorithm, but turns out it doesn't do some of the "basic" stuff the old one did.

We've got a few paths to remediate these cases:
- We can solve them "in post" in the same way that Capture Context "fixed" otherwise ambiguous or meh results from the old search
- We can update the way we're building blocks
- We can update the way we're searching blocks

I think each of these cases will need something a little different. For the case where I want to return the line missing a "do", where we're grabbing the correct "end", I think we can/should fix that in post.

There's a lot of other cases where I'm not quite sure.

Also worth mentioning that some of these failure modes may change if any values are placed "inside" them for example:

```
def foo
  def blerg
end
```

Returns a match saying the problem is `def foo` but if you put something "in" the `def blerg` it says the problem is `def blerg`

```
def foo
  def blerg
    print lol
end
```

Let's take it one step at a time, focus on one failure case, see if it breaks other cases when it's fixed...then iterate.