WIP Initial BlockNodeContext
Fixes 6 out of 8 of the prior failures. The concept is quite simple as well. If a node leans one way, capture in the direction needed to theoretically fix it until we find a node that is leaning.

It fails though when we look at a larger example because blocks tend to self balance for valid code and this produces very large outputs for large inputs.

It's not suitable for a final answer, but it's a good first step. I introduced a failing test with "handles heredocs" to capture that need.