Update block building
We were making some questionable decisions when building blocks. I refactored out the IndentTree building so that it can be done step by step so we can observe it. This allows for tests that are quite verbose, but should allow for use of intuition "that shouldn't be the next step" as blocks are being built.

I found that some nodes at the same indent were being expanded to cover the entire "inner" before other blocks could capture them. For example in:

```
        describe "things" do
          it "blerg" do
            print foo1
          end # one

          it "flerg"
            print foo2
          end # two

          it "zlerg" do
            print foo3
          end # three
        end
```

This block:

```

          it "zlerg" do
            print foo3
          end # three

```

Would expand to capture the above `end # two` even though logically it should be captured by the `print foo2` block.

To accommodate this I added an extra condition to block node expansion saying to not capture a node leaning in the opposite direction as expansion if that node is also a leaf. Essentially if a node is leaning right, it needs to expand up. I want either it to expand itself, or for it's "inner" equal block to expand out to capture it.

The other case I added was to enforce the indentation check for nodes leaning opposite of expansion direction, but only on first expansion (other wise we end up with "split" blocks where one block will not engulf all inner blocks.

Those changes exposed a problem with the `next_indent` calculation where sometimes it would come back at a higher value than the current indent which is not correct. Fixing this by adding a final check/guarantee when deriving that value.

This change seems good in isolation but is causing a lot of test failures due to tight coupling between tests and implementation. I need to go back and re-work the tests to see if there's any fundamental "disagreements" or if they just need to be updated to new/better values.

There's one problem fundamental to the :both case that seems not well handled here