Respect SyntaxError#detailed_message `highlight` kwarg
The kwarg `highlight` is intended to enable/disable control characters https://twitter.com/schneems/status/1532693550474924032. This change brings DeadEnd into alignment with error_highlight with the ability to turn on/off the control characters at will.

One downside of this implementation is that for the case where highlighting is enabled I'm calling super twice. This might have performance implications if someone else is overwriting SyntaxError and doing something possibly expensive (much like DeadEnd is doing).

Alternatively we could try to detect or strip the control code. This would be harder and possibly error prone.

Also addresses:

- Disable/enable dead end based on kwarg https://github.com/ruby/ruby/pull/5516/files#diff-f2e6c18de8171e389aeff0b8d25b2755253108a59a13baf7167c2ffa0af36a58R1388-R1390