Drop support for old 3.2.0 preview versions
Ruby 3.2.0rc1 introduced SyntaxError#path in addition to the ability to monkey patch SyntaxError. That means all code inside of this syntax error is guaranteed to be on Ruby 3.2. As the contents of the preview release were unstable, we do not have to continue to support older preview releases that did not contain SyntaxError#path