Allow testing SyntaxError#detailed_message
Since syntax_suggest is now integrated with Ruby as a default gem it is now a bit tougher to test. The default gem monkeypatches SyntaxError so when it's monkey patched again via the gem the result will be different.

This change moves the module into a method that can be used to prepend any class for testing so we can assert behavior against our code in isolation while integration tests assert with the full stack.