Refactor logic, add docs, add tests
AroundBlockScan became a dumping ground for logic as it was where the scanner lived. Since moving the scanner to its own class it makes more sense to move this logic out of AroundBlockScan.

In the future I want to move the one-off methods out of CaptureCodeContext into their own classes and have them individually tested.